feat: add the CodeEditor.vue component to centrally manage UI modules.
